* commit '757dbf80d81bfdd06affe2b8fcfdc1eab9e3b418':
  BEN-1073 Address review comments
  BEN-1073 Minor cleanup of toolchain file
  BEN-1073 Fixed minor bug in UnifiedCudaArray
  BEN-1073 This commit does two major things:
  BEN-1073 Restored some debug sanity checking, removed previously due to compilation error that snuck into develop

+++ src/basecaller/basecaller/traceAnalysis/HostHFMetricsFilter.cpp
@@ -38,7 +38,7 @@ void HostHFMetricsFilter::FinalizeBlock()
 {
     for (size_t l = 0; l < lanesPerBatch_; ++l)
     {
-        metrics_[l].FinalizeMetrics(realtimeActivityLabels_, frameRate_);
+        metrics_[l].FinalizeMetrics(realtimeActivityLabels_, static_cast<float>(frameRate_));
     }
 }
